medagent_env/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-medagent_env"
version = "0.1.0"
description = "Medical Clinical Assessment environment for OpenEnv (Scaler Round 1)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.115.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "build>=1.2.2",
    "pytest>=8.0.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["medagent_env", "medagent_env.server"]
package-dir = { "medagent_env" = ".", "medagent_env.server" = "server" }
